+++ test_unstructured_ingest/test-ingest-biomed-path.sh
@@ -1,22 +1,10 @@
 #!/usr/bin/env bash
 # shellcheck disable=SC2317
-# NOTE(crag): remove above shellcheck line when the biomed issue is fixed
-echo "Skipping test-ingest-biomed-path.sh,"
-echo "see https://github.com/Unstructured-IO/unstructured/issues/468"
-echo
-exit 0
 
 set -e
 
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 cd "$SCRIPT_DIR"/.. || exit 1
-
-if [[ "$CI" == "true" ]]; then
-    if [ "$(( RANDOM % 10))" -lt 1 ] ; then
-        echo "Skipping ingest 90% of biomed tests to avoid the occaisonal ftp issue."
-        exit 0
-    fi
-fi
 
 if [[ "$(find test_unstructured_ingest/expected-structured-output/biomed-ingest-output-path/ -type f -size +10k | wc -l)" != 1 ]]; then
     echo "The test fixtures in test_unstructured_ingest/expected-structured-output/biomed-ingest-output-path/ look suspicious. At least one of the files is too small."
@@ -27,6 +15,9 @@ fi
 PYTHONPATH=. ./unstructured/ingest/main.py \
     --metadata-exclude filename,file_directory,metadata.data_source.date_processed \
     --biomed-path "oa_pdf/07/07/sbaa031.073.PMC7234218.pdf" \
+    --biomed-max-retries 5 \
+    --biomed-max-request-time 30 \
+    --biomed-decay .3 \
     --structured-output-dir biomed-ingest-output-path \
     --num-processes 2 \
     --partition-strategy hi_res \